+++ front/src/components/organisms/CategoriesList.tsx
@@ -63,6 +63,7 @@ const categoryImg: Record<string, string> = {
   React: icons.react, // TODO: Add React Icon
   Spring: icons.spring, // TODO: Add Spring Icon
   Laravel: icons.laravel, // TODO: Add Spring Icon
+  BBDD:icons.bbdd, // TODO: Add BBDD Icon
 }
 
 export const CategoriesList: FC = () => {