+++ sso/src/routes/routes.ts
@@ -3,5 +3,6 @@ export const pathRoot = {
     auth: '/api/v1/auth',
     users: '/api/v1/users',
     tokens: '/api/v1/tokens',
+    itinerary: '/api/v1/itinerary',
   },
 }